Update console connection tools
1. Refactor code
2. Support interactive_shell
