+++ tests/common/connections/ssh_console_conn.py
@@ -1,7 +1,6 @@
-from tests.common.connections.base_console_conn import CONSOLE_SSH
 import time
 import re
-from base_console_conn import BaseConsoleConn
+from base_console_conn import BaseConsoleConn, CONSOLE_SSH
 from netmiko.ssh_exception import NetMikoAuthenticationException
 
 class SSHConsoleConn(BaseConsoleConn):
